# docs/archive/product-history/LobsterPark_Docs_V1_4/LobsterPark_OpenAPI_V1_4.yaml
openapi: 3.1.0
info:
  title: Lobster Park API
  version: 1.4.0
  description: >
    OpenAPI contract for Lobster Park. V1.4 adds Secrets CRUD endpoints,
    enhances alerts/notifications filters, fixes PatchInstanceRequest.name
    maxLength, and adds quotaJson to Tenant schema.
servers:
- url: https://lobster-park.internal
security:
- cookieAuth: []
paths:
...
  /api/v1/platform/settings:
    get:
      summary: List platform settings
      operationId: listPlatformSettings
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopePagedPlatformSettings'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/platform/settings/{settingKey}:
    get:
      summary: Get platform setting
      operationId: getPlatformSetting
      parameters:
      - $ref: '#/components/parameters/SettingKey'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopePlatformSetting'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    put:
      summary: Update platform setting
      operationId: putPlatformSetting
      parameters:
      - $ref: '#/components/parameters/SettingKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - settingValueJson
              properties:
                settingValueJson:
                  $ref: '#/components/schemas/AnyJsonValue'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopePlatformSetting'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/platform/schemas/{runtimeVersion}:
    get:
      summary: Get JSON schema for a runtime version
      operationId: getRuntimeSchema
      parameters:
      - $ref: '#/components/parameters/RuntimeVersion'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: lp_access
  parameters:
...
    SettingKey:
      in: path
      name: settingKey
      required: true
      schema:
        type: string
      description: Platform setting key
    RuntimeVersion:
      in: path
      name: runtimeVersion
      required: true
      schema:
        type: string
      description: Approved OpenClaw runtime version
    # ── V1.4 NEW ──
...
  schemas:
    EnvelopeBase:
      type: object
      properties:
        requestId:
          type: string
        code:
          type: integer
        message:
          type: string
...
    EnvelopeError:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            type: object
            additionalProperties: true
...
    PlatformSetting:
      type: object
      properties:
        id:
          type: string
        settingKey:
          type: string
        settingValueJson:
          $ref: '#/components/schemas/AnyJsonValue'
        description:
          type: string
        updatedBy:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    AnyJsonValue:
      description: Any valid JSON value used by platform settings.
      oneOf:
      - type: object
        additionalProperties: true
      - type: array
        items: {}
      - type: string
      - type: number
      - type: boolean
      - type: 'null'
    # ── V1.4 NEW: Secrets schemas ─────────────────────────────────
...
    EnvelopePagedPlatformSettings:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            type: object
            properties:
              items:
                type: array
                items:
                  $ref: '#/components/schemas/PlatformSetting'
    EnvelopePlatformSetting:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/PlatformSetting'
...
  responses:
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EnvelopeError'
...
  headers:
    RetryAfter:
      description: Seconds until the client may retry
      schema:
        type: integer